+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend_user_service_functions.go
@@ -215,6 +215,36 @@ func (backend *DockerKurtosisBackend) GetUserServiceLogs(
 	return successfulUserServicesLogs, erroredUserServices, nil
 }
 
+func (backend *DockerKurtosisBackend) PauseService(
+	ctx context.Context,
+	enclaveId enclave.EnclaveID,
+	serviceId service.ServiceGUID) error {
+	containerId, _, err := backend.getSingleUserService(ctx, enclaveId, serviceId)
+	if err != nil {
+		return stacktrace.Propagate(err, "Failed to get information about service '%v' from Kurtosis backend.", serviceId)
+	}
+	err = backend.dockerManager.PauseContainer(ctx, containerId)
+	if err != nil {
+		return stacktrace.Propagate(err, "Failed to pause service' %v' running in container '%v'", serviceId, containerId)
+	}
+	return nil
+}
+
+func (backend *DockerKurtosisBackend) UnpauseService(
+	ctx context.Context,
+	enclaveId enclave.EnclaveID,
+	serviceId service.ServiceGUID) error {
+	containerId, _, err := backend.getSingleUserService(ctx, enclaveId, serviceId)
+	if err != nil {
+		return stacktrace.Propagate(err, "Failed to get information about service '%v' from Kurtosis backend.", serviceId)
+	}
+	err = backend.dockerManager.UnpauseContainer(ctx, containerId)
+	if err != nil {
+		return stacktrace.Propagate(err, "Failed to unpause service '%v' running in container '%v'", serviceId, containerId)
+	}
+	return nil
+}
+
 func (backend *DockerKurtosisBackend) RunUserServiceExecCommands(
 	ctx context.Context,
 	enclaveId enclave.EnclaveID,